Extract the Node-safe transaction export pipeline
`TransactionExportActions.tsx` was a five-hundred-line thunk that did four
separable jobs: fill in historical fiat values, render CSV, render QBO, and
render the Bitwave format. Only the last step needed React Native, and only
for writing the file.

`fillTxsFiat` asks the rates server what each transaction was worth on the day
it happened, which is the part that makes an export more than a dump of native
amounts. `txExport/format` renders the three formats. `exportTxInfo` holds the
Bitwave account mapping the exporter needs.

The thunk keeps the file-writing and the share sheet, and calls the same
renderers. `TransactionsExportScene` follows it. A test now covers `fillTxsFiat`
against a partial wallet, which is all it reads.

The formats matter here: an export that a person reconciles against their books
has to be byte-identical whichever tool produced it, and the only way to be
sure of that is for one renderer to produce both.
